initial movment of files
