+++ package.json
@@ -11,7 +11,11 @@
       "**/react-native-*",
       "**/react-native-*/**",
       "**/rn-*",
-      "**/rn-*/**"
+      "**/rn-*/**",
+      "**/react-navigation",
+      "**/react-navigation/**",
+      "**/react-navigation-*",
+      "**/react-navigation-*/**"
     ]
   },
   "scripts": {
