+++ packages/apollos-ui-auth/.npmignore
@@ -1,0 +1,2 @@
+*
+!lib/**
