+++ packages/apollos-ui-auth/README.md
@@ -1,0 +1,3 @@
+# Apollos UI: Auth
+
+React Native components to log your users in.
