+++ packages/apollos-ui-auth/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@apollosproject/ui-passes",
+  "name": "@apollosproject/ui-auth",
   "version": "0.8.0-alpha.8",
   "homepage": "https://github.com/ApollosProject/apollos-prototype#readme",
   "main": "src/index.js",
@@ -24,9 +24,6 @@
     "testMatch": [
       "<rootDir>/src/**/?(*.)(test|tests).{js,jsx,mjs}"
     ],
-    "setupFiles": [
-      "./jest.setup.js"
-    ],
     "modulePaths": [
       "<rootDir>/src/"
     ],
@@ -37,14 +34,16 @@
   "peerDependencies": {
     "@apollos-project/ui-kit": "*",
     "react": "*",
-    "react-native": "*",
-    "react-native-passkit-wallet": "*",
-    "rn-fetch-blob": "*"
+    "react-native": "*"
   },
   "dependencies": {
     "@apollosproject/ui-kit": "^0.8.0-alpha.8",
-    "graphql-tag": "^2.10.1",
-    "react-apollo": "^2.4.1"
+    "apollo-client": "2.4.2",
+    "graphql-tag": "2.9.2",
+    "lodash": "^4.17.11",
+    "react-apollo": "2.2.4",
+    "react-navigation": "2.11.2",
+    "react-native-keyboard-aware-scroll-view": "0.7.0"
   },
   "devDependencies": {
     "@apollosproject/babel-preset-apollos": "^0.8.0-alpha.8",
@@ -56,12 +55,7 @@
     "react": "^16.8.2",
     "react-dom": "16.6.3",
     "react-native": "0.57.5",
-    "react-native-linear-gradient": "2.4.0",
-    "react-native-passkit-wallet": "^0.1.4",
-    "react-native-svg": "6.5.2",
-    "react-navigation": "2.11.2",
-    "react-test-renderer": "16.6.1",
-    "rn-fetch-blob": "^0.10.15"
+    "react-test-renderer": "16.6.1"
   },
   "publishConfig": {
     "access": "public"
