+++ packages/apollos-ui-auth/src/ProtectedTouchable.js
