+++ packages/apollos-ui-auth/src/getLoginState.js
