+++ packages/apollos-ui-auth/src/login/authenticate.js
