+++ packages/apollos-ui-auth/src/logout.js
