+++ packages/apolloschurchapp/package.json
@@ -66,6 +66,7 @@
   "dependencies": {
     "@apollosproject/ui-kit": "^0.8.0-alpha.8",
     "@apollosproject/ui-passes": "^0.8.0-alpha.8",
+    "@apollosproject/ui-auth": "^0.8.0-alpha.8",
     "apollo-cache-inmemory": "1.3.5",
     "apollo-cache-persist": "0.1.1",
     "apollo-client": "2.4.2",
@@ -114,6 +115,7 @@
     "react-native-video-controls": "2.2.3",
     "react-native-webview": "^5.2.0",
     "react-navigation": "2.11.2",
+    "react-navigation-tabs": "0.6.0",
     "recompose": "^0.30.0",
     "rn-fetch-blob": "^0.10.15",
     "rn-placeholder": "1.2.0",
