+++ packages/apolloschurchapp/rn-cli.config.js
@@ -4,6 +4,7 @@ const blacklist = require('metro-config/src/defaults/blacklist');
 const localDeps = [
   path.resolve('..', 'apollos-ui-kit'),
   path.resolve('..', 'apollos-ui-passes'),
+  path.resolve('..', 'apollos-ui-auth'),
 ];
 
 module.exports = {
