+++ packages/apolloschurchapp/src/Providers.js
@@ -1,5 +1,6 @@
 import { nest } from 'recompose';
 import { Providers } from '@apollosproject/ui-kit';
+import { Provider as AuthProvider } from '@apollosproject/ui-auth';
 import ClientProvider from './client';
 
-export default nest(ClientProvider, Providers);
+export default nest(ClientProvider, AuthProvider, Providers);
