+++ packages/apolloschurchapp/src/store/index.js
@@ -1,24 +1,19 @@
 import { merge, get } from 'lodash';
 import gql from 'graphql-tag';
-import getLoginState from 'apolloschurchapp/src/auth/getLoginState';
 import { track, events } from 'apolloschurchapp/src/analytics';
 import { client, CACHE_LOADED } from "../client"; // eslint-disable-line
 import updatePushId from '../notifications/updatePushId';
-import getAuthToken from './getAuthToken';
 // TODO: this will require more organization...ie...not keeping everything in one file.
 // But this is simple while our needs our small.
 
 export const schema = `
   type Query {
-    authToken: String
     mediaPlayer: MediaPlayerState
-    isLoggedIn: Boolean
     devicePushId: String
     cacheLoaded: Boolean
   }
 
   type Mutation {
-    logout
     mediaPlayerUpdateState(isPlaying: Boolean, isFullscreen: Boolean, isVisible: Boolean): Boolean
     mediaPlayerSetPlayhead(currentTime: Float): Boolean
     mediaPlayerPlayNow(
@@ -31,9 +26,6 @@ export const schema = `
     ): Boolean
 
     cacheMarkLoaded
-
-    handleLogin(authToken: String!)
-
     updateDevicePushId(pushId: String!)
   }
 
@@ -65,7 +57,6 @@ export const schema = `
 
 export const defaults = {
   __typename: 'Query',
-  authToken: null,
   cacheLoaded: false,
   pushId: null,
   mediaPlayer: {
@@ -83,58 +74,8 @@ export const defaults = {
 let trackId = 0;
 
 export const resolvers = {
-  Query: {
-    isLoggedIn: () => {
-      // When logging out, this query returns an error.
-      // Rescue the error, and return false.
-      try {
-        const { authToken } = client.readQuery({ query: getAuthToken });
-        return !!authToken;
-      } catch (e) {
-        return false;
-      }
-    },
-  },
+  Query: {},
   Mutation: {
-    logout: () => {
-      client.resetStore();
-      track({ eventName: events.UserLogout });
-      return null;
-    },
-
-    handleLogin: async (root, { authToken }, { cache }) => {
-      try {
-        await cache.writeQuery({
-          query: getAuthToken,
-          data: { authToken },
-        });
-        await cache.writeQuery({
-          query: getLoginState,
-          data: { isLoggedIn: true },
-        });
-        await cache.writeData({
-          data: { authToken },
-        });
-
-        const { pushId } = cache.readQuery({
-          query: gql`
-            query {
-              pushId
-            }
-          `,
-        });
-        if (pushId) {
-          updatePushId({ pushId });
-        }
-
-        track({ eventName: events.UserLogin });
-      } catch (e) {
-        throw e.message;
-      }
-
-      return null;
-    },
-
     mediaPlayerPlayNow: (root, trackInfo, { cache }) => {
       const query = gql`
         query {
